requirements.txt
selenium==4.15.2
python-dotenv==1.0.0
webdriver-manager==4.0.1
requests==2.31.0
pillow==10.1.0
schedule==1.2.0
